Clean up repository file
