+++ postdetails/src/main/java/com/neaniesoft/vermilion/postdetails/data/PrettyTimeModule.kt
@@ -1,0 +1,14 @@
+package com.neaniesoft.vermilion.postdetails.data
+
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import org.ocpsoft.prettytime.PrettyTime
+
+@Module
+@InstallIn(SingletonComponent::class)
+class PrettyTimeModule {
+    @Provides
+    fun providePrettyTime(): PrettyTime = PrettyTime()
+}
